requirement.txt
gitpython
inquirerpy
rich